# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GTCP — Gas Trading & Commercial Platform API
  description: |
    **Gastrans Serbia · Backend REST API**

    Система управления газовыми транзитными потоками, договорами мощности, биллингом и аукционами.

    ### Нормативная база
    - **CAM NC EU 2017/459** — механизмы распределения мощности
    - **АЕРС 05-145** — официальные тарифы Gastrans
    - **Gastrans Network Code** — NC Art. 5 (Credit Support), Art. 6.3.1 (EXIT-SERBIA),
      Art. 8.3 (Surrender), Art. 18 (Fuel Gas), Art. 20 (Billing & Late Payment)
    - **MAR0277-24** — расписание аукционов CAM NC 2025-2026 (ENTSOG)

    ### Аутентификация
    Все endpoints (кроме `/auth/login`) требуют JWT Bearer token.

    ### Lifecycle аукциона
    ```
    Free Capacity → POST /auctions/bids (DRAFT)
                  → POST /auctions/bids/{id}/submit (SUBMITTED)
                  → POST /auctions/bids/{id}/result (WON/LOST)
                  → POST /auctions/bids/{id}/create-contract (CONTRACT_CREATED)
                  → POST /billing (Invoice)
    ```
  version: 5.0.0
  contact:
    name: GTCP Backend
    email: krav4ukleo@gmail.com
  license:
    name: Proprietary — Gastrans Serbia
    url: https://gastrans.rs
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development
  - url: https://gtcp.gastrans.rs/api/v1
    description: Production (Sprint 6)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token, срок действия 24 часа.
        Получить: `POST /auth/login` → `{ token }`.
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: secret123 }
    LoginResponse:
      type: object
      properties:
        token:        { type: string, description: JWT access token (24h) }
        refreshToken: { type: string, description: JWT refresh token (7d) }
        user:
          type: object
          properties:
            id:       { type: integer }
            username: { type: string }
            role:     { type: string, enum: [ADMIN, OPERATOR, TRADER, VIEWER, FINANCE] }

    # ─── Shipper ──────────────────────────────────────────────
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Получить JWT токен
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
            example:
              username: admin
              password: admin123
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          description: Неверные учётные данные
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновить access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Новый access token
...